Delivery

+++ src/main/java/com/generation/javeat/utils/Utils.java
@@ -4,19 +4,19 @@ import com.generation.javeat.model.entities.Restaurant;
 
 public class Utils {
 
-    static public double calculateDistanceToRestaurant(Restaurant restaurant, Integer userX, Integer userY) {
+    static public Integer calculateDistanceToRestaurant(Restaurant restaurant, Integer userX, Integer userY) {
         // Coordinate del ristorante
-        int restaurantX = restaurant.getPositionX();
-        int restaurantY = restaurant.getPositionY();
+        Integer restaurantX = restaurant.getPositionX();
+        Integer restaurantY = restaurant.getPositionY();
         // Calcola la differenza tra le coordinate x e y
-        int diffX = restaurantX - userX;
+        Integer diffX = restaurantX - userX;
         int diffY = restaurantY - userY;
         // Calcola la somma dei quadrati delle differenze
-        int squaredDiffX = diffX * diffX;
-        int squaredDiffY = diffY * diffY;
+        Integer squaredDiffX = diffX * diffX;
+        Integer squaredDiffY = diffY * diffY;
         // Calcola la radice quadrata della somma dei quadrati delle differenze
         double distance = Math.sqrt(squaredDiffX + squaredDiffY);
-        return distance;
+        return (int)distance;
     }
 
     public static boolean isValidPassword(String password) {
